# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OM Pay API",
        "description": "API pour l'application OM Pay - Gestion des paiements et transferts",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Serveur de développement"
        },
        {
            "url": "https://om-pay-mk4h.onrender.com/api",
            "description": "Serveur de production"
        }
    ],
    "paths": {
...
        "/{numeroCompte}/portefeuille/solde": {
            "get": {
                "tags": [
                    "Portefeuille"
                ],
                "summary": "Consulter le solde",
                "description": "Récupère le solde actuel du portefeuille OM Pay",
                "operationId": "49e8248ac091dc36f681590419a4ae4d",
                "parameters": [
                    {
                        "name": "numeroCompte",
                        "in": "path",
                        "description": "Numéro de compte de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "7735434534"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Solde récupéré avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "solde": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 15000.5
                                                },
                                                "devise": {
                                                    "type": "string",
                                                    "example": "XOF"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Token invalide"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/{numeroCompte}/portefeuille/transactions": {
            "get": {
                "tags": [
                    "Portefeuille"
                ],
                "summary": "Historique des transactions",
                "description": "Récupère l'historique des transactions avec filtres optionnels",
                "operationId": "ea98ec36dc94d4d4e55c46007ac7171a",
                "parameters": [
                    {
                        "name": "numeroCompte",
                        "in": "path",
                        "description": "Numéro de compte de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "7735434534"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numéro de page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "limite",
                        "in": "query",
                        "description": "Nombre d'éléments par page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Type de transaction (transfert, paiement, etc.)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "dateDebut",
                        "in": "query",
                        "description": "Date de début (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "dateFin",
                        "in": "query",
                        "description": "Date de fin (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "statut",
                        "in": "query",
                        "description": "Statut de la transaction",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Historique récupéré avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "transactions": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "idTransaction": {
                                                                "type": "string",
                                                                "example": "01HN1234567890ABCDEF"
                                                            },
                                                            "type": {
                                                                "type": "string",
                                                                "example": "transfert"
                                                            },
                                                            "montant": {
                                                                "description": "Montant avec signe +/-",
                                                                "type": "string",
                                                                "example": "-5000"
                                                            },
                                                            "montantNumerique": {
                                                                "description": "Montant numérique sans signe",
                                                                "type": "number",
                                                                "format": "float",
                                                                "example": 5000
                                                            },
                                                            "devise": {
                                                                "type": "string",
                                                                "example": "XOF"
                                                            },
                                                            "typeOperation": {
                                                                "type": "string",
                                                                "enum": [
                                                                    "debit",
                                                                    "credit"
                                                                ],
                                                                "example": "debit"
                                                            },
                                                            "expediteur": {
                                                                "properties": {
                                                                    "numeroTelephone": {
                                                                        "type": "string",
                                                                        "example": "771234567"
                                                                    },
                                                                    "nom": {
                                                                        "type": "string",
                                                                        "example": "John Doe"
                                                                    }
                                                                },
                                                                "type": "object",
                                                                "nullable": true
                                                            },
                                                            "destinataire": {
                                                                "properties": {
                                                                    "numeroTelephone": {
                                                                        "type": "string",
                                                                        "example": "781234567"
                                                                    },
                                                                    "nom": {
                                                                        "type": "string",
                                                                        "example": "Jane Smith"
                                                                    }
                                                                },
                                                                "type": "object",
                                                                "nullable": true
                                                            },
                                                            "marchand": {
                                                                "properties": {
                                                                    "nom": {
                                                                        "type": "string",
                                                                        "example": "Boutique Express"
                                                                    },
                                                                    "categorie": {
                                                                        "type": "string",
                                                                        "example": "Alimentation"
                                                                    }
                                                                },
                                                                "type": "object",
                                                                "nullable": true
                                                            },
                                                            "statut": {
                                                                "type": "string",
                                                                "example": "reussi"
                                                            },
                                                            "dateTransaction": {
                                                                "type": "string",
                                                                "format": "date-time"
                                                            },
                                                            "reference": {
                                                                "type": "string",
                                                                "example": "OM20251111131953ABC123"
                                                            },
                                                            "frais": {
                                                                "type": "number",
                                                                "format": "float",
                                                                "example": 0
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "pagination": {
                                                    "properties": {
                                                        "pageActuelle": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "totalPages": {
                                                            "type": "integer",
                                                            "example": 5
                                                        },
                                                        "totalElements": {
                                                            "type": "integer",
                                                            "example": 100
                                                        },
                                                        "elementsParPage": {
                                                            "type": "integer",
                                                            "example": 20
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/{numeroCompte}/portefeuille/transactions/{idTransaction}": {
            "get": {
                "tags": [
                    "Portefeuille"
                ],
                "summary": "Détails d'une transaction",
                "description": "Récupère les détails d'une transaction spécifique",
                "operationId": "5defd98fcc63ca7fd7921a85667cbf69",
                "parameters": [
                    {
                        "name": "numeroCompte",
                        "in": "path",
                        "description": "Numéro de compte de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "7735434534"
                        }
                    },
                    {
                        "name": "idTransaction",
                        "in": "path",
                        "description": "ID de la transaction",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de la transaction récupérés",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "string"
                                                },
                                                "montant": {
                                                    "type": "number",
                                                    "format": "float"
                                                },
                                                "type": {
                                                    "type": "string"
                                                },
                                                "statut": {
                                                    "type": "string"
                                                },
                                                "date": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Transaction non trouvée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Transaction non trouvée"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}